extract formatter into trait
