+++ src/Livewire/Widgets/RecurringRevenueForecast.php
@@ -8,11 +8,11 @@ use FluxErp\Livewire\Dashboard\Dashboard;
 use FluxErp\Livewire\Order\OrderList;
 use FluxErp\Livewire\Support\Widgets\Charts\BarChart;
 use FluxErp\Models\Pivots\OrderSchedule;
+use FluxErp\Traits\HasTemporalXAxisFormatter;
 use FluxErp\Traits\Livewire\IsTimeFrameAwareWidget;
 use FluxErp\Traits\MoneyChartFormattingTrait;
 use FluxErp\Traits\Widgetable;
 use Illuminate\Database\Eloquent\Builder;
-use Livewire\Attributes\Js;
 use Livewire\Attributes\Locked;
 use Livewire\Attributes\Renderless;
 use Livewire\Livewire;
@@ -21,7 +21,7 @@ use Throwable;
 
 class RecurringRevenueForecast extends BarChart implements HasWidgetOptions
 {
-    use IsTimeFrameAwareWidget, MoneyChartFormattingTrait, Widgetable;
+    use HasTemporalXAxisFormatter, IsTimeFrameAwareWidget, MoneyChartFormattingTrait, Widgetable;
 
     public ?array $chart = [
         'type' => 'bar',
@@ -175,23 +175,5 @@ class RecurringRevenueForecast extends BarChart implements HasWidgetOptions
     public function showTitle(): bool
     {
         return true;
-    }
-
-    #[Js]
-    public function xAxisFormatter(): string
-    {
-        return <<<'JS'
-            let name;
-            if (typeof val === 'string' && val.includes('->')) {
-                name = val.split('->')[1];
-                val = val.split('->')[0];
-            }
-
-            if (/^\d{4}$/.test(val)) {  // Value with 4 digits is a year
-                return val + (name ? ' (' + name + ')' : '');
-            }
-
-            return new Date(val).toLocaleDateString(document.documentElement.lang) + (name ? ' (' + name + ')' : '');
-        JS;
     }
 }
