+++ src/Livewire/Widgets/TotalOrdersCount.php
@@ -10,17 +10,17 @@ use FluxErp\Livewire\Support\Widgets\Charts\LineChart;
 use FluxErp\Models\Order;
 use FluxErp\Support\Metrics\Charts\Line;
 use FluxErp\Support\Metrics\Value;
+use FluxErp\Traits\HasTemporalXAxisFormatter;
 use FluxErp\Traits\Livewire\IsTimeFrameAwareWidget;
 use FluxErp\Traits\Widgetable;
 use Illuminate\Database\Eloquent\Builder;
-use Livewire\Attributes\Js;
 use Livewire\Attributes\Renderless;
 use Livewire\Livewire;
 use TeamNiftyGmbH\DataTable\Helpers\SessionFilter;
 
 class TotalOrdersCount extends LineChart implements HasWidgetOptions
 {
-    use IsTimeFrameAwareWidget, Widgetable;
+    use HasTemporalXAxisFormatter, IsTimeFrameAwareWidget, Widgetable;
 
     public static function dashboardComponent(): array|string
     {
@@ -124,24 +124,6 @@ class TotalOrdersCount extends LineChart implements HasWidgetOptions
             ->store();
 
         $this->redirectRoute('orders.orders', navigate: true);
-    }
-
-    #[Js]
-    public function xAxisFormatter(): string
-    {
-        return <<<'JS'
-            let name;
-            if (typeof val === 'string' && val.includes('->')) {
-                name = val.split('->')[1];
-                val = val.split('->')[0];
-            }
-
-            if (/^\d{4}$/.test(val)) {  // Value with 4 digits is a year
-                return val + (name ? ' (' + name + ')' : '');
-            }
-
-            return new Date(val).toLocaleDateString(document.documentElement.lang) + (name ? ' (' + name + ')' : '');
-        JS;
     }
 
     protected function getListeners(): array
