+++ src/Livewire/Widgets/TotalRevenue.php
@@ -10,18 +10,18 @@ use FluxErp\Livewire\Support\Widgets\Charts\LineChart;
 use FluxErp\Models\Order;
 use FluxErp\Support\Metrics\Charts\Line;
 use FluxErp\Support\Metrics\Value;
+use FluxErp\Traits\HasTemporalXAxisFormatter;
 use FluxErp\Traits\Livewire\IsTimeFrameAwareWidget;
 use FluxErp\Traits\MoneyChartFormattingTrait;
 use FluxErp\Traits\Widgetable;
 use Illuminate\Database\Eloquent\Builder;
-use Livewire\Attributes\Js;
 use Livewire\Attributes\Renderless;
 use Livewire\Livewire;
 use TeamNiftyGmbH\DataTable\Helpers\SessionFilter;
 
 class TotalRevenue extends LineChart implements HasWidgetOptions
 {
-    use IsTimeFrameAwareWidget, MoneyChartFormattingTrait, Widgetable;
+    use HasTemporalXAxisFormatter, IsTimeFrameAwareWidget, MoneyChartFormattingTrait, Widgetable;
 
     public static function dashboardComponent(): array|string
     {
@@ -131,24 +131,6 @@ class TotalRevenue extends LineChart implements HasWidgetOptions
     public function showTitle(): bool
     {
         return true;
-    }
-
-    #[Js]
-    public function xAxisFormatter(): string
-    {
-        return <<<'JS'
-            let name;
-            if (typeof val === 'string' && val.includes('->')) {
-                name = val.split('->')[1];
-                val = val.split('->')[0];
-            }
-
-            if (/^\d{4}$/.test(val)) {  // Value with 4 digits is a year
-                return val + (name ? ' (' + name + ')' : '');
-            }
-
-            return new Date(val).toLocaleDateString(document.documentElement.lang) + (name ? ' (' + name + ')' : '');
-        JS;
     }
 
     protected function getListeners(): array
